proj/requirements.txt
api==0.0.7
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Django==2.1.5
django-filter==2.1.0
django-jalali-date==0.2.9
django-money==0.14.4
django-rest-auth==0.9.3
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
ez-setup==0.9
idna==2.8
jdatetime==3.2.0
jdcal==1.4
Markdown==3.0.1
nose==1.3.7
openpyxl==2.6.0
Pillow==5.4.1
psycopg2==2.7.7
pur==5.2.1
py-moneyed==0.8.0
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
restaurant==0.1.0
six==1.12.0
soupsieve==1.7.3
urllib3==1.24.1
